+++ src/core/handle_get.c
@@ -1,7 +1,7 @@
 /*
  *
- * Copyright (C) 2019-2021, Broadband Forum
- * Copyright (C) 2016-2021  CommScope, Inc
+ * Copyright (C) 2019-2022, Broadband Forum
+ * Copyright (C) 2016-2022  CommScope, Inc
  *
  * Redistribution and use in source and binary forms, with or without
  * modification, are permitted provided that the following conditions
@@ -57,7 +57,6 @@
 // Structure used to marshall entries in get group vector for a path expression
 typedef struct
 {
-    int separator_split;    // Used to split resolved parameter path into resolved object and resolved sub-path. It is a count of number of dot separators included in the 'object' portion of the path
     int index;              // start index of parameters in get group vector for this path expression
     int num_entries;        // number of entries in the get group vector for this path expression
     int err_code;           // error code if path resolution failed for this path expression
@@ -66,9 +65,9 @@ typedef struct
 
 //------------------------------------------------------------------------------
 // Forward declarations. Note these are not static, because we need them in the symbol table for USP_LOG_Callstack() to show them
-void ExpandGetPathExpression(int get_expr_index, char *path_expr, get_expr_info_t *gi, group_get_vector_t *ggv);
+void ExpandGetPathExpression(int get_expr_index, char *path_expr, int depth, get_expr_info_t *gi, group_get_vector_t *ggv);
 void FormPathExprResponse(int get_expr_index, char *path_expr, get_expr_info_t *gi, group_get_vector_t *ggv, Usp__Msg *resp);
-void AddResolvedPathResult(Usp__GetResp__RequestedPathResult *req_path_result, char *path, char *value, int separator_split);
+void AddResolvedPathResult(Usp__GetResp__RequestedPathResult *req_path_result, char *path, char *value);
 Usp__GetResp__ResolvedPathResult *FindResolvedPath(Usp__GetResp__RequestedPathResult *req_path_result, char *obj_path);
 Usp__Msg *CreateGetResp(char *msg_id);
 Usp__GetResp__RequestedPathResult *AddGetResp_ReqPathRes(Usp__Msg *resp, char *requested_path, int err_code, char *err_msg);
@@ -97,6 +96,7 @@ void MSG_HANDLER_HandleGet(Usp__Msg *usp, char *controller_endpoint, mtp_reply_t
     int num_path_expr;
     Usp__Msg *resp = NULL;
     int size;
+    int depth;
     group_get_vector_t ggv;
     get_expr_info_t *get_expr_info;
 
@@ -123,6 +123,13 @@ void MSG_HANDLER_HandleGet(Usp__Msg *usp, char *controller_endpoint, mtp_reply_t
         goto exit;
     }
 
+    // Calculate the number of hierarchical levels to traverse in the data model when performing partial path resolution
+    depth = usp->body->request->get->max_depth;
+    if (depth == 0)
+    {
+        depth = FULL_DEPTH;
+    }
+
     // Allocate vector to store marshalling info for each path expression
     size = num_path_expr*sizeof(get_expr_info_t);
     get_expr_info = USP_MALLOC(size);
@@ -132,7 +139,7 @@ void MSG_HANDLER_HandleGet(Usp__Msg *usp, char *controller_endpoint, mtp_reply_t
     GROUP_GET_VECTOR_Init(&ggv);
     for (i=0; i < num_path_expr; i++)
     {
-        ExpandGetPathExpression(i, path_exprs[i], &get_expr_info[i], &ggv);
+        ExpandGetPathExpression(i, path_exprs[i], depth, &get_expr_info[i], &ggv);
     }
 
     // Get all parameters
@@ -166,13 +173,14 @@ exit:
 **
 ** \param   get_expr_index - index of the path expression in the USP Get request message
 ** \param   path_expr - USP path expression specifying which parameters to get
+** \param   depth - Number of hierarchical levels to traverse in the data model when performing partial path resolution
 ** \param   gi - pointer to info about specified path expression
 ** \param   ggv - pointer to group get vector to add params found in this path expression to
 **
 ** \return  None
 **
 **************************************************************************/
-void ExpandGetPathExpression(int get_expr_index, char *path_expr, get_expr_info_t *gi, group_get_vector_t *ggv)
+void ExpandGetPathExpression(int get_expr_index, char *path_expr, int depth, get_expr_info_t *gi, group_get_vector_t *ggv)
 {
     int err;
     str_vector_t params;
@@ -184,7 +192,7 @@ void ExpandGetPathExpression(int get_expr_index, char *path_expr, get_expr_info_
     STR_VECTOR_Init(&params);
     INT_VECTOR_Init(&group_ids);
     MSG_HANDLER_GetMsgRole(&combined_role);
-    err = PATH_RESOLVER_ResolveDevicePath(path_expr, &params, &group_ids, kResolveOp_Get, &gi->separator_split, &combined_role, 0);
+    err = PATH_RESOLVER_ResolveDevicePath(path_expr, &params, &group_ids, kResolveOp_Get, depth, &combined_role, 0);
     if (err != USP_ERR_OK)
     {
         gi->err_code = err;
@@ -270,13 +278,8 @@ void FormPathExprResponse(int get_expr_index, char *path_expr, get_expr_info_t *
     {
         gge = &ggv->vector[gi->index + i];
 
-#ifdef GET_RESPONSE_SIMPLE_FORMAT
         // Simple format contains a resolved_path_result for every object (and sub object)
-        AddResolvedPathResult(req_path_result, gge->path, gge->value, 0);
-#else
-        // Original format contains a resolved_path_result from the first point in the path expression that is resolved
-        AddResolvedPathResult(req_path_result, gge->path, gge->value, gi->separator_split);
-#endif
+        AddResolvedPathResult(req_path_result, gge->path, gge->value);
     }
 }
 
@@ -291,21 +294,18 @@ void FormPathExprResponse(int get_expr_index, char *path_expr, get_expr_info_t *
 ** \param   req_path_result - pointer to requested_path_result to add this entry to
 ** \param   path - full data model path of the parameter
 ** \param   value - value of the parameter
-** \param   separator_split - denotes where to split the parameter path based on the number of separators for the object that required resolution
-**                            The path is split into an object (that required resolution),
-**                            and a sub path which did not require resolution
 **
 ** \return  None
 **
 **************************************************************************/
-void AddResolvedPathResult(Usp__GetResp__RequestedPathResult *req_path_result, char *path, char *value, int separator_split)
+void AddResolvedPathResult(Usp__GetResp__RequestedPathResult *req_path_result, char *path, char *value)
 {
     char obj_path[MAX_DM_PATH];
     char *param_name;
     Usp__GetResp__ResolvedPathResult *resolved_path_res;
 
     // Split the parameter into the parent object path and the name of the parameter within the object
-    param_name = TEXT_UTILS_SplitPathAtSeparator(path, obj_path, sizeof(obj_path), separator_split);
+    param_name = TEXT_UTILS_SplitPath(path, obj_path, sizeof(obj_path));
 
     // Add a resolved path result, if we don't already have one for the specified parent object
     resolved_path_res = FindResolvedPath(req_path_result, obj_path);